Implementation of the RFP solver.
- Axioms for `rfp.sub` and `rfp.div`.
- Linear approximation lemma for `rfp.round`.
- Support for the property terms.
- Bug fixes.

+++ src/util/real_floatingpoint.h
@@ -47,7 +47,7 @@ inline static Integer maxExponent(uint32_t size)
  */
 inline static Integer minExponent(uint32_t size)
 {
-  return Integer(Integer(2) - maxExponent(size));
+  return Integer(Integer(1) - maxExponent(size));
 }
 
 /** Get the maximum normal value.
